+++ packages/react/src/utils/useOutboundClickListener.tsx
@@ -3,7 +3,19 @@ import { MatomoInstance } from '../types'
 
 const useOutboundClickListener = (matomoInstance: MatomoInstance): void => {
   const handleOutboundClick = (event: MouseEvent) => {
-    const { target } = event
+    // The target is not guaranteed to be a link, it could be a child element.
+    // Look up the element's parent until the anchor element is found.
+    const findLinkElement = (el: EventTarget | null): HTMLElement | null => {
+      if (el instanceof HTMLAnchorElement && el.href) {
+        return el
+      }
+      if (el instanceof HTMLElement && el.parentElement) {
+        return findLinkElement(el.parentElement)
+      }
+      return null
+    }
+
+    const target = findLinkElement(event.target)
 
     if (!(target instanceof HTMLAnchorElement)) {
       return